Feature: task related endpoints with user access control via jwt
* Add field check for json input in group

* Add task controller

* Change each task to store a user id

* Add recently completed task endpoint

* Remove unused dependencies

* Fix errors from merging

* Add placeholder to function

* Add login user control to user task endpoints

* Fix recent task endpoint

* Add json serialization to tasks

+++ app/services/task_service.py
@@ -1,16 +1,25 @@
 from datetime import timedelta
 from typing import Optional
+
+from bson import ObjectId
 from app.models.task_model import Task
 from app.repositories.task_repository import TaskRepository
+from app.repositories.user_repository import UserRepository
 from app.utils.time import get_current_time
 
 
 class TaskService:
-    def __init__(self, task_repository: TaskRepository) -> None:
+    def __init__(
+        self, task_repository: TaskRepository, user_repository: UserRepository
+    ) -> None:
         self.task_repository = task_repository
+        self.user_repository = user_repository
 
-    def create_task(self, name: str, description: str = "") -> Task:
-        result = self.task_repository.create_task(name, description)
+    def create_task(self, user_id: str, name: str, description: str = "") -> Task:
+        user = self.user_repository.find_by_id(user_id)
+        if user is None:
+            raise ValueError("User does not exist")
+        result = self.task_repository.create_task(user_id, name, description)
         if result is None:
             raise ValueError("Task could not be created")
         return result
@@ -21,13 +30,21 @@ class TaskService:
             raise ValueError("Task does not exist")
         return task
 
+    def find_task_by_user_id(self, user_id: str) -> list[Task]:
+        return self.task_repository.find_tasks_by_user_id(user_id)
+
     def update_task(
         self,
+        current_user_id: str,
         task_id: str,
         name: Optional[str] = None,
         description: Optional[str] = None,
         completed: Optional[bool] = None,
     ) -> None:
+        task = self.find_task_by_id(task_id)
+        if ObjectId(task.user_id) != ObjectId(current_user_id):
+            raise ValueError("User does not have permission to update task")
+
         update_fields = {}
         if name:
             update_fields["name"] = name
@@ -46,7 +63,10 @@ class TaskService:
         if update_result.modified_count == 0:
             raise ValueError("Task could not be updated")
 
-    def delete_task(self, task_id: str) -> None:
+    def delete_task(self, current_user_id: str, task_id: str) -> None:
+        task = self.find_task_by_id(task_id)
+        if ObjectId(task.user_id) != ObjectId(current_user_id):
+            raise ValueError("User does not have permission to delete task")
         delete_result = self.task_repository.delete_task_by_id(task_id)
         if delete_result.deleted_count == 0:
             raise ValueError("Task could not be deleted")
